+++ src/main/java/com/charles/librarymgt/controllers/BookController.java
@@ -2,69 +2,47 @@ package com.charles.librarymgt.controllers;
 
 import com.charles.librarymgt.dtos.BookDto;
 import com.charles.librarymgt.models.Book;
-import com.charles.librarymgt.repository.BookRepository;
+import com.charles.librarymgt.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.PagedModel;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 
 @RestController
 @RequestMapping("/api")
 public class BookController {
 
-    @Autowired
-    private BookRepository bookRepository;
 
     @Autowired
-    private PagedResourcesAssembler<BookDto> pagedResourcesAssembler;
+    private BookService bookService;
 
     @GetMapping("/books")
     PagedModel<EntityModel<BookDto>> index(
             @PageableDefault(page = 0, size = Integer.MAX_VALUE, sort = {"publicationYear"}) Pageable paging) {
 
-        Page<Book> result = bookRepository.findAll(paging);
-
-        return pagedResourcesAssembler.toModel(result.map(BookDto::new));
+        return bookService.getAllBooks(paging);
     }
 
     @GetMapping("/books/{id}")
     BookDto show(@PathVariable Long id) {
-        return new BookDto(bookRepository.findById(id).get());
+        return bookService.getBook(id);
     }
 
     @PostMapping("/books")
     BookDto store(@RequestBody Book book) {
-        return new BookDto(bookRepository.save(book));
+        return bookService.addBook(book);
     }
 
     @PutMapping("/books/{id}")
     BookDto update(@RequestBody Book newBook, @PathVariable Long id) {
-        var book = bookRepository.findById(id).get();
-
-        if (!book.getTitle().equals(newBook.getTitle())) {
-            book.setTitle(newBook.getTitle());
-        }
-        if (!book.getAuthor().equals(newBook.getAuthor())) {
-            book.setAuthor(newBook.getAuthor());
-        }
-        if (!book.getIsbn().equals(newBook.getIsbn())) {
-            book.setIsbn(newBook.getIsbn());
-        }
-        if (!book.getPublicationYear().equals(newBook.getPublicationYear())) {
-            book.setPublicationYear(newBook.getPublicationYear());
-        }
-
-        return new BookDto(bookRepository.save(book));
+        return bookService.updateBook(newBook, id);
     }
 
     @DeleteMapping("/books/{id}")
     void delete(@PathVariable Long id) {
-        bookRepository.deleteById(id);
+        bookService.deleteBook(id);
     }
 }
