+++ src/main/java/com/charles/librarymgt/repository/BorrowBookRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BorrowBookRepository extends JpaRepository<BorrowRecord, Long> {
 
-    BorrowRecord findBorrowRecordByBookAndPatron(Book book, Patron patron);
+    BorrowRecord findBorrowRecordByBookAndPatronAndReturnDateNull(Book book, Patron patron);
 }
